增加了对ip的拉黑功能

+++ qu-pojo/src/main/java/com/mjc/entity/SysLog.java
@@ -12,6 +12,6 @@ public class SysLog {
     private String method;  //请求方法
     private String params;  //请求参数
     private String ip;      //操作ip
-    private Long time;      //执行时长
+    private Double time;      //执行时长
     private LocalDateTime createTime;  //操作时间
 }
